+++ src/main/java/com/platzimarket/persistence/crud/ProductoCrudRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 public interface ProductoCrudRepository extends CrudRepository<Producto, Integer> {
     List<Producto> findByIdCategoriaOrderByNombreAsc(int idCategoria);
     Optional<List<Producto>> findByStockLessThanAndEstado(int stock, boolean estado);
-    List<Producto> findByPrecioVentaOrderByNombreAsc(int precioVenta);
+    List<Producto> findByPrecioVentaOrderByNombreAsc(double precioVenta);
 }
